pyproject.toml
[tool.poetry]
name = "animoapi"
version = "0.1.0"
description = ""
authors = ["preitychib <preeti.chib2015@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
djangorestframework = "^3.14.0"
Django = "^4.1.3"
environs = "^9.5.0"

[tool.poetry.dev-dependencies]
yapf = "^0.32.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
